12/13/16 committing lesson 11 Individual assignment (print Resource Report)

+++ HungerGames/src/byui/cit260/hungerGames/model/Resource.java
@@ -35,7 +35,7 @@ public enum Resource implements Serializable{
         return description;
     }
 
-    public double getHealthValue() {
+    public int getHealthValue() {
         return healthValue;
     }
